Extract all other methods into separated files
