+++ helpers.py
@@ -1,6 +1,7 @@
-from datetime import datetime
+from telegram import InlineKeyboardMarkup
 
-from dateutil.tz import tzlocal
+from logger import log_print
+from random_content import random_content
 
 
 def start(bot, update):
@@ -24,9 +25,38 @@ def bug(bot, update):
                      parse_mode='markdown')
 
 
-def log_print(message, *username):
-    timestamp = datetime.now(tzlocal()).strftime("[%d/%b/%Y:%H:%M:%S %z]")
-    if username:
-        print("{0}: {1} by @{2}".format(timestamp, message, "".join(username)))
-    else:
-        print("{0}: {1}".format(timestamp, message))
+def hat(bot, update):
+    user_id = update.message.from_user.id
+    faculties = ['Gryffindor', 'Slytherin', 'Hufflepuff', 'Ravenclaw']
+    faculty = faculties[user_id % len(faculties)]
+
+    bot.send_message(chat_id=update.message.chat_id,
+                     text='*{0}*'.format(faculty),
+                     reply_to_message_id=update.message.message_id,
+                     parse_mode='markdown')
+    log_print('Hat answer is {0} for user_id = {1}'.format(faculty, user_id))
+
+
+def chat_id(bot, update):
+    chat_id = update.message.chat_id
+    username = update.message.from_user.username
+    bot.send_message(chat_id=chat_id,
+                     text="`{0}`".format(chat_id),
+                     reply_to_message_id=update.message.message_id,
+                     parse_mode='markdown')
+    log_print('Chat id {0}'.format(chat_id), username)
+
+
+def buttons(bot, update):
+    query = update.callback_query
+    keyboard = InlineKeyboardMarkup([[]])
+    bot.edit_message_reply_markup(chat_id=query.message.chat_id,
+                                  message_id=query.message.message_id,
+                                  reply_markup=keyboard)
+    if query.data in ['cat_1', 'cat_5',
+                      'dog_1', 'dog_5',
+                      'ibash_1', 'ibash_5',
+                      'loglist_1', 'loglist_5']:
+        command, value = query.data.split('_')
+        query.message.text = '/{}'.format(command)
+        random_content(bot, query, [value])
